# src-tauri/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Fono Local Transcription API",
    "version": "1.0.0",
    "description": "Authenticated loopback-only API exposed by the Fono desktop app."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:17832",
      "description": "Default local address; override with FONO_API_PORT before starting Fono."
    }
  ],
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/v1/health": {
      "get": {
        "summary": "Read local service health",
        "responses": {
          "200": { "description": "Ready", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Health" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/v1/transcriptions": {
      "post": {
        "summary": "Upload audio and create an asynchronous transcription job",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["audio", "model"],
                "properties": {
                  "audio": { "type": "string", "format": "binary", "description": "WAV, MP3, FLAC, OGG/Vorbis or OGG/Opus; maximum 100 MiB." },
                  "language": { "type": "string", "default": "auto", "maxLength": 64 },
                  "model": { "type": "string", "enum": ["tiny", "base", "small", "medium", "large", "large_turbo"], "description": "Required identifier of the model currently selected in Fono." }
                }
              }
            }
          }
        },
        "responses": {
          "202": { "description": "Queued", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Job" } } } },
          "400": { "$ref": "#/components/responses/InvalidUpload" },
          "413": { "$ref": "#/components/responses/FileTooLarge" },
          "415": { "$ref": "#/components/responses/UnsupportedAudio" },
          "422": { "$ref": "#/components/responses/InvalidAudio" },
          "429": { "$ref": "#/components/responses/QueueFull" }
        }
      }
    },
    "/v1/transcription-jobs": {
      "post": {
        "summary": "Create a job from normalized PCM",
        "description": "Low-level compatibility endpoint. Prefer /v1/transcriptions for files.",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TranscriptionRequest" } } }
        },
        "responses": {
          "202": { "description": "Queued", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Job" } } } },
          "429": { "$ref": "#/components/responses/QueueFull" }
        }
      }
    },
    "/v1/transcription-jobs/{id}": {
      "get": {
        "summary": "Read job state",
        "parameters": [{ "$ref": "#/components/parameters/JobId" }],
        "responses": {
          "200": { "description": "Job", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Job" } } } },
          "404": { "$ref": "#/components/responses/JobNotFound" }
        }
      }
    },
    "/v1/transcription-jobs/{id}/cancel": {
      "post": {
        "summary": "Cancel a queued or running job",
        "parameters": [{ "$ref": "#/components/parameters/JobId" }],
        "responses": {
          "200": { "description": "Cancelled job", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Job" } } } },
          "404": { "$ref": "#/components/responses/JobNotFound" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer" } },
    "parameters": { "JobId": { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } } },
    "responses": {
      "Unauthorized": { "description": "Missing or invalid bearer token" },
      "InvalidUpload": { "description": "Malformed multipart request" },
      "FileTooLarge": { "description": "Audio exceeds the 100 MiB limit" },
      "UnsupportedAudio": { "description": "Only WAV, MP3, FLAC, OGG/Vorbis and OGG/Opus are supported" },
      "InvalidAudio": { "description": "Corrupted audio or duration over the configured limit" },
      "QueueFull": { "description": "Retry after an active job finishes" },
      "JobNotFound": { "description": "No job with this id exists" }
    },
    "schemas": {
      "Health": { "type": "object", "required": ["protocol_version", "state"], "properties": { "protocol_version": { "type": "integer", "example": 1 }, "state": { "type": "string", "example": "ready" } } },
      "TranscriptionRequest": { "type": "object", "required": ["pcm_samples", "language", "model"], "properties": { "pcm_samples": { "type": "array", "items": { "type": "integer", "format": "int16" }, "description": "Mono signed PCM at 16 kHz." }, "language": { "type": "string", "example": "auto" }, "model": { "type": "string", "enum": ["tiny", "base", "small", "medium", "large", "large_turbo"], "example": "base", "description": "Must match Fono's selected model." } } },
      "Result": { "type": "object", "required": ["protocol_version", "text", "model"], "properties": { "protocol_version": { "type": "integer" }, "text": { "type": "string" }, "detected_language": { "type": "string", "nullable": true }, "audio_seconds": { "type": "number", "nullable": true }, "transcribe_seconds": { "type": "number", "nullable": true }, "model": { "type": "string" }, "backend": { "type": "string", "nullable": true } } },
      "Job": { "type": "object", "required": ["id", "state", "created_at_ms"], "properties": { "id": { "type": "string", "example": "tr_0000000000000001" }, "state": { "type": "string", "enum": ["queued", "preparing", "transcribing", "completed", "failed", "cancelled"] }, "created_at_ms": { "type": "integer", "format": "int64", "description": "Unix epoch timestamp in milliseconds." }, "started_at_ms": { "type": "integer", "format": "int64", "nullable": true }, "finished_at_ms": { "type": "integer", "format": "int64", "nullable": true }, "result": { "$ref": "#/components/schemas/Result", "nullable": true }, "error": { "nullable": true } } }
    }
  }
}